feat: Move Header and Hero components to server-side
This commit transitions the Header and Hero section components from
client-side to server-side rendering. This change helps to avoid Flash
of Unstyled Content (FOUC) or unnecessary loading spinners,
resulting in a more seamless user experience.

Furthermore, custom hooks utilized within the Header and Hero
section components have been reorganized into a new directory, /hooks/ui,
for better code structure and maintainability. Previously, these hooks
were located in the /hooks/misc directory.
