+++ lib/stateLogics/hooks/misc.tsx
@@ -1,13 +1,13 @@
 import { PATH_APP } from '@constAssertions/data';
 import { Labels, Todos } from '@lib/types';
 import { selectorSessionLabels } from '@states/atomEffects/labels';
-import { selectorSessionTodoItem, atomSelectorTodoItem } from '@states/atomEffects/todos';
+import { atomSelectorTodoItem, selectorSessionTodoItem } from '@states/atomEffects/todos';
 import { atomLabelNew, atomSelectorLabelItem } from '@states/labels';
 import { atomTodoModalMini, atomTodoModalOpen } from '@states/modals';
 import { atomTodoNew, selectorFilterTodoIdsByPathname } from '@states/todos';
 import equal from 'fast-deep-equal/react';
 import { useRouter } from 'next/router';
-import { RefObject, useCallback, useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo } from 'react';
 import { RecoilState, RecoilValue, useRecoilCallback, useRecoilValue } from 'recoil';
 
 export const useFilterTodoIdsWithPathname = () => {
@@ -111,77 +111,6 @@ export const useNextQuery = ({ path, key }: { path?: string; key?: string }) => 
   return value;
 };
 
-export const useHorizontalScrollPosition = (ref: RefObject<HTMLDivElement>) => {
-  const [leftPosition, setLeftPosition] = useState(-1);
-  const [rightPosition, setRightPosition] = useState(-1);
-  const [isOverflow, setIsOverflow] = useState(false);
-
-  const initialOverflown = ref.current && ref.current.scrollWidth > 300;
-
-  useEffect(() => {
-    const currentRef = ref.current;
-
-    if (currentRef) {
-      const overScrollWidth = currentRef.clientWidth < currentRef.scrollWidth;
-      const overflown = overScrollWidth || (initialOverflown as boolean);
-      setIsOverflow(overflown);
-    }
-
-    const handleScroll = () => {
-      if (currentRef) {
-        const scrollWidth = currentRef.scrollWidth - currentRef.clientWidth;
-        const rightPosition = scrollWidth && scrollWidth - currentRef.scrollLeft;
-        setLeftPosition(ref.current.scrollLeft);
-        setRightPosition(rightPosition as number);
-      }
-    };
-
-    ref.current && ref.current.addEventListener('scroll', handleScroll);
-    return () => {
-      currentRef && currentRef.removeEventListener('scroll', handleScroll);
-    };
-  }, [initialOverflown, ref]);
-
-  return { leftPosition, rightPosition, isOverflow };
-};
-
-export const useWindowWidth = () => {
-  const getWindowWidth = () => {
-    return window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth;
-  };
-  const [windowWidth, setWindowWidth] = useState(getWindowWidth());
-
-  const handleResize = useCallback(() => {
-    setWindowWidth(getWindowWidth());
-  }, []);
-
-  useEffect(() => {
-    window.addEventListener('resize', handleResize);
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, [handleResize]);
-
-  return windowWidth;
-};
-
-export const useVerticalScrollPosition = () => {
-  const [scrollPosition, setScrollPosition] = useState(0);
-
-  const handleScroll = () => {
-    const position = window.pageYOffset;
-    setScrollPosition(position);
-  };
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll);
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, []);
-
-  return scrollPosition;
-};
 
 export const useCompareToQueryLabels = () => {
   return useRecoilCallback(({ snapshot }) => (compare: Labels[]) => {
