+++ pages/index.tsx
@@ -2,12 +2,8 @@ import { LayoutHome } from '@layouts/home';
 import dynamic from 'next/dynamic';
 import { Fragment, ReactElement } from 'react';
 
-const Hero = dynamic(() => import('@components/sections/hero').then((mod) => mod.Hero), {
-  ssr: false,
-});
-const Header = dynamic(() => import('@components/sections/header').then((mod) => mod.Header), {
-  ssr: false,
-});
+const Hero = dynamic(() => import('@components/sections/hero').then((mod) => mod.Hero));
+const Header = dynamic(() => import('@components/sections/header').then((mod) => mod.Header));
 
 const Home = () => {
   return (
